update vnpay payment

+++ src/main/java/com/api_coffee_store/api_coffee_store/dtos/request/ProductVariantRequest.java
@@ -9,6 +9,6 @@ import lombok.*;
 @AllArgsConstructor
 public class ProductVariantRequest {
     private String size;
-    private Double price;
+    private long price;
     private boolean status;
 }
